normal lighting not quite working

+++ GLUT interlacer/OpenGL with 3D/TangentWaveFrontPolygon.h
@@ -5,7 +5,7 @@
 class TangentWaveFrontPolygon : public WaveFrontPolygon
 {
 public:
-	std::vector<double> tangents;
+	std::vector<float> tangents;
 
 	TangentWaveFrontPolygon(std::string objFile);
 };
